More control vars and refactor utils
